// src/state/types.ts
export type SerializedBigNumber = string
// ...
export interface Vault extends VaultConfig {
  lpTotalInQuoteToken?: SerializedBigNumber
  tokenPriceVsQuote?: SerializedBigNumber
  tokenAmountTotal?: SerializedBigNumber
  quoteTokenAmountTotal?: SerializedBigNumber
  lpTotalSupply?: SerializedBigNumber
  tags: string[]
  userData?: {
    allowance: string
    tokenBalance: string
    stakedBalance: string
    earnings: string
  }
  sushiData?: {
    // set of variables used to calculate apr for vaults that use the SushiSwap Strategy
    poolDailyRewardUsd: SerializedBigNumber
    SushiLpTotalInQuoteToken: SerializedBigNumber
  }
  quickData?: {
    dailyReward: SerializedBigNumber
    QuickLpTotalInQuoteToken?: SerializedBigNumber
  }
  apeData?: {
    poolDailyRewardUsd: SerializedBigNumber
    apeLpTotalInQuoteToken: SerializedBigNumber
  }
  spadeData?: {
    spadePriceUsd: SerializedBigNumber
    spadePerBlock: SerializedBigNumber
    spadePoolWeight: SerializedBigNumber
    spadeLpTotalInQuoteToken: SerializedBigNumber
  }
  bcharityData?: {
    pid: number
  }
}